# docs/api/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This document describes the RESTful APIs provided or assumed by the\nKubernetes Application Runtime (KAR). It consists of two sets of APIs:\n+ The **sidecar** APIs are implemented by the KAR sidecar.\n+ The **application component** APIs are meant to be implemented by application components.\n\nThe **sidecar** APIs are divided into logical groups:\n+ **Actors**: APIs to invoke actor methods.\n+ **Services**: APIs to invoke service endpoints.\n+ **Callbacks**: APIs to await the response to an asynchronous actor or service invocation.\n+ **Events**: APIs to publish to event sinks or subscribe actors to event sources.\n+ **Reminders**: APIs to schedule future actor invocations.\n+ **State**: APIs to manage the persistent state of actors.\n+ **System**: APIs for controlling the KAR runtime mesh.\n\nThe **application component** APIs are divided into logical groups:\n+ **Actor runtime**: APIs invoked by the sidecar to manage actor instances\nhosted by the application component.",
    "title": "KAR",
    "version": "v1"
  },
  "basePath": "/kar",
  "paths": {
...
    "/impl/v1/actor/{actorType}/{actorId}": {
      "get": {
        "description": "### Allocate the language-level state for the specified actor instance\n\nCauses the language-level actor runtime to allocate storage for the\nactor instance and to invoke its initialization/activation method\nif one is provided.  This operation must be successfully completed\nbefore any POST operations on this actor instance are performed.",
        "schemes": [
          "http"
        ],
        "tags": [
          "actor-runtime"
        ],
        "summary": "actor allocation",
        "operationId": "idImplActorGet",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      },
      "delete": {
        "description": "### Deallocate the language-level state for the specified actor instance\n\nThe optional passivation/deactivation method for the actor type will\nbe invoked.  After it completes, the language-level storate for the actor\ninstance will be released. After this operation is invoked, no more POST\noperations on this actor instance may be performed unless a GET is first\nperformed to re-initialize the language-level state of the actor instance.",
        "schemes": [
          "http"
        ],
        "tags": [
          "actor-runtime"
        ],
        "summary": "actor deallocation",
        "operationId": "idImplActorDelete",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      }
    },
    "/impl/v1/actor/{actorType}/{actorId}/{methodName}": {
      "post": {
        "description": "### Invoke an actor method of the specified actor instance\n\nInvokes the actor method on the actor instance within the session specified in the path.\nThe body of the request will contain the actual paramters on which to invoke the method.",
        "consumes": [
          "application/kar+json"
        ],
        "produces": [
          "application/kar+json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "actor-runtime"
        ],
        "summary": "actor invocation",
        "operationId": "idImplActorPost",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ActorType",
            "description": "The actor type",
            "name": "actorType",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ActorID",
            "description": "The actor instance id",
            "name": "actorId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "example": "computeMyResult",
            "x-go-name": "MethodName",
            "description": "The actor method to be invoked",
            "name": "methodName",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Session",
            "description": "Optionally specific the session to use when performing the call.  Enables re-entrancy for nested actor calls and tracking parent-child relationships",
            "name": "session",
            "in": "query"
          },
          {
            "example": "[3, 'hello', { msg: 'Greetings' }]",
            "description": "A possibly empty array containing the arguments with which to invoke the target actor method.",
            "name": "ActorMethodArguments",
            "in": "body",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/response200CallActorResult"
          },
          "202": {
            "$ref": "#/responses/response202"
          },
          "204": {
            "$ref": "#/responses/response204ActorNoContentResult"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      }
    },
    "/v1/actor/{actorType}/{actorId}": {
      "delete": {
        "description": "### Completely remove an actor instance\n\nAll user-level and runtime state of the actor instance indicated by\n`actorType` and `actorId` will be asynchronously deleted.",
        "schemes": [
          "http"
        ],
        "tags": [
          "actors"
        ],
        "summary": "actor",
        "operationId": "idActorDelete",
        "responses": {
          "202": {
            "$ref": "#/responses/response202"
          },
          "404": {
            "$ref": "#/responses/response404"
          },
          "500": {
            "$ref": "#/responses/response500"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "actorCallResult": {
      "type": "object",
      "title": "actorCallResult encodes the result of invoking an actor method.",
      "properties": {
        "error": {
          "description": "If true, indicates that the method execution resulted in an error/exception",
          "type": "boolean",
          "x-go-name": "Error"
        },
        "message": {
          "description": "When error is true, the error message",
          "type": "string",
          "x-go-name": "Message"
        },
        "stack": {
          "description": "When error is true, the stack trace for the error",
          "type": "string",
          "x-go-name": "Stack"
        },
        "tailCall": {
          "description": "If true, indicates that the Value represents a continuation",
          "type": "boolean",
          "x-go-name": "TailCall"
        },
        "value": {
          "description": "The value being returned as the result of the method",
          "type": "object",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/IBM/kar/core/internal/runtime"
    },
...
  },
  "responses": {
    "response200": {
      "description": "A success message.",
      "headers": {
        "body": {
          "type": "string",
          "example": "OK",
          "description": "A success message"
        }
      }
    },
    "response200CallActorResult": {
      "description": "The result of invoking the actor method",
      "schema": {
        "$ref": "#/definitions/actorCallResult"
      },
      "headers": {
        "Body": {
          "description": "The result returned by the actor method"
        }
      }
    },
...
    "response202": {
      "description": "An asynchronous operation has been accepted for eventual execution"
    },
...
    "response204ActorNoContentResult": {
      "description": "Indicates that the result of an actor method is void or undefined."
    },
...
    "response404": {
      "description": "Response indicating requested resource is not found",
      "headers": {
        "body": {
          "type": "string",
          "example": "Not Found",
          "description": "Requested resource is not found"
        }
      }
    },
    "response500": {
      "description": "A message describing the error",
      "headers": {
        "body": {
          "type": "string",
          "example": "Internal Server Error",
          "description": "A message describing the error"
        }
      }
    },
...
  },
...
}